requirements.txt
langgraph==0.2.28
langchain-openai>=0.1.25
langchain-community>=0.2.10
langsmith>=0.1.77
tavily-python==0.3.3
yfinance>=0.2.40
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
jupyter
ipykernel
